Fix codeclimate issues and synchronize offer when closed
When closed the view is updated on every app showing it.
