+++ app/src/androidTest/java/ch/epfl/sweng/swenggolf/CreateUserActivityTest.java
@@ -20,7 +20,6 @@ import ch.epfl.sweng.swenggolf.profile.User;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static android.support.test.espresso.action.ViewActions.replaceText;
 import static android.support.test.espresso.action.ViewActions.scrollTo;
 import static android.support.test.espresso.action.ViewActions.typeText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
