+++ app/src/main/java/ch/epfl/sweng/swenggolf/database/FakeDatabase.java
@@ -3,25 +3,17 @@ package ch.epfl.sweng.swenggolf.database;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.TreeMap;
-
-import ch.epfl.sweng.swenggolf.offer.Category;
-import ch.epfl.sweng.swenggolf.offer.Offer;
 
 /**
  * Mocked Database which works locally.
  */
-public class FakeDatabase extends Database {
+public class FakeDatabase extends DatabaseListHandler {
     private final Map<String, Object> database;
     private final Map<String, List<ValueListener>> listeners;
     private Set<String> workingOnEntry;
@@ -41,10 +33,6 @@ public class FakeDatabase extends Database {
         this.workingOnEntry = new HashSet<>();
     }
 
-    private static void handleError(String attribute) {
-        throw new IllegalArgumentException("The attribute " + attribute + " doesn't exist");
-    }
-
     /**
      * Creates a Database already filled with users and offers.
      *
@@ -54,66 +42,45 @@ public class FakeDatabase extends Database {
         return new FilledFakeDatabase();
     }
 
-    private static String getGetter(String attribute) {
-        return "get" + Character.toUpperCase(attribute.charAt(0))
-                + attribute.substring(1, attribute.length());
-    }
-
-    @NonNull
-    private static <T> Comparator<T> getComparator(final Method method) {
-        return new Comparator<T>() {
-            @Override
-            public int compare(T o1, T o2) {
-                Object attribute1 = invokeGetter(method, o1);
-                Object attribute2 = invokeGetter(method, o2);
-
-                return compareAttributes(attribute1, attribute2);
-            }
-        };
-    }
-
-    private static <T> Object invokeGetter(Method method, T invokedOn) {
-        try {
-            return method.invoke(invokedOn);
-        } catch (IllegalAccessException e) {
-            throw new IllegalArgumentException("Can't access the attribute");
-        } catch (InvocationTargetException e) {
-            throw new IllegalArgumentException("Cannot call method on generic parameter T");
-        }
-    }
-
-    private static <T> int compareAttributes(T attribute1, T attribute2) {
-        if (attribute1 instanceof Comparable && attribute2 instanceof Comparable) {
-            return ((Comparable) attribute1).compareTo(attribute2);
-        }
-        throw new IllegalArgumentException("The attribute is not comparable");
-    }
-
     @Override
     public void write(@NonNull String path, @NonNull String id, @NonNull Object object) {
         if (working) {
-            path = path + "/" + id;
             database.put(path, object);
-            String[] listenningToChildren = path.split("/");
-            notifyChildren(listenningToChildren);
-            if(listeners.containsKey(path)) {
-                for(ValueListener listener : listeners.get(path)) {
-                    listener.onDataChange(object);
-                }
+            path = path + "/" + id;
+            String[] listeningToChildren = path.split("/");
+            notifyChildren(listeningToChildren);
+            notifyPathListeners(path, object);
+        }
+    }
+
+    private void notifyPathListeners(String path, Object object) {
+        if (listeners.containsKey(path)) {
+            List<ValueListener> objectListeners = listeners.get(path);
+            for (ValueListener listener : objectListeners) {
+                listener.onDataChange(object);
             }
         }
     }
 
+    private String[] createChildrenPaths(String[] listenningToChildren) {
+        String[] paths = new String[listenningToChildren.length];
+        StringBuilder currentPath = new StringBuilder();
+        for (int i=0;i<paths.length;++i) {
+            paths[i] = currentPath.toString();
+            currentPath.append("/").append(listenningToChildren[i]);
+        }
+        return paths;
+    }
+
     private void notifyChildren(String[] listenningToChildren) {
-        StringBuilder currentLevel = new StringBuilder();
+        listenningToChildren = createChildrenPaths(listenningToChildren);
         for (String level : listenningToChildren) {
-            String current = currentLevel.toString();
-            if(listeners.containsKey(current)) {
-                for (ValueListener listener : listeners.get(current)) {
-                    listener.onDataChange(getList(current));
+            if (listeners.containsKey(level)) {
+                List<ValueListener> childrenListeners = listeners.get(level);
+                for (ValueListener listener : childrenListeners) {
+                    listener.onDataChange(getList(level));
                 }
             }
-            currentLevel.append("/").append(level);
         }
     }
 
@@ -148,7 +115,7 @@ public class FakeDatabase extends Database {
     public <T> void listen(@NonNull String path, @NonNull String id,
                            @NonNull ValueListener<T> listener, @NonNull Class<T> c) {
         String pathToListen = path + "/" + id;
-        if(listeners.containsKey(pathToListen)) {
+        if (listeners.containsKey(pathToListen)) {
             listeners.get(pathToListen).add(listener);
         } else {
             List<ValueListener> pathListeners = new ArrayList<>();
@@ -162,7 +129,7 @@ public class FakeDatabase extends Database {
     public <T> void deafen(@NonNull String path, @NonNull String id,
                            @NonNull ValueListener<T> listener) {
         path = path + "/" + id;
-        if(listeners.containsKey(path)) {
+        if (listeners.containsKey(path)) {
             listeners.get(path).remove(listener);
         }
     }
@@ -171,134 +138,24 @@ public class FakeDatabase extends Database {
         return working && !workingOnEntry.contains(key);
     }
 
-    @Override
-    public <T> void readList(@NonNull String path, @NonNull ValueListener<List<T>> listener,
-                             @NonNull Class<T> c) {
-        if (working) {
-            List<T> list = getList(path);
-            listener.onDataChange(list);
-        } else {
-            listener.onCancelled(DbError.UNKNOWN_ERROR);
-        }
-    }
 
-    @Override
-    public <T> void readList(@NonNull String path, @NonNull ValueListener<List<T>> listener,
-                             @NonNull Class<T> c, AttributeFilter filter) {
-        if (working) {
-
-            List<T> list = getList(path);
-
-            List<T> newList = filterList(c, filter.getAttribute(), filter.getValue(), list);
-            listener.onDataChange(newList);
-        } else {
-            listener.onCancelled(DbError.UNKNOWN_ERROR);
-        }
-    }
-
-    @Override
-    public <T> void readList(@NonNull String path, @NonNull ValueListener<List<T>> listener,
-                             @NonNull Class<T> c, AttributeOrdering ordering) {
-        if (working) {
-            List<T> unsortedList = getList(path);
-            List<T> list = sortList(c, ordering, unsortedList);
-            listener.onDataChange(list);
-        } else {
-            listener.onCancelled(DbError.UNKNOWN_ERROR);
-        }
-    }
-
-    @NonNull
-    private <T> List<T> sortList(@NonNull Class<T> c, AttributeOrdering ordering,
-                                 List<T> unsortedList) {
-        final Method method;
-        String getterName = getGetter(ordering.getAttribute());
-        try {
-            method = c.getDeclaredMethod(getterName);
-        } catch (NoSuchMethodException e) {
-            throw new IllegalArgumentException("The getter for the attribute does not exist");
-        }
-        Comparator<T> comparator = getComparator(method);
-        Collections.sort(unsortedList, comparator);
-        if (ordering.isDescending()) {
-            Collections.reverse(unsortedList);
-        }
-        int minSize = Math.min(ordering.getNumberOfElements(), unsortedList.size());
-        return unsortedList.subList(0, minSize);
-    }
-
-    private <T> List<T> filterList(@NonNull Class<T> c, String attribute, String value,
-                                   List<T> list) {
-        List<T> filtered = new ArrayList<>();
-        try {
-
-            //Use reflection to check the attribute
-            Method method = c.getDeclaredMethod(getGetter(attribute));
-            filtered = createFilteredList(method, list, value);
-
-        } catch (IllegalAccessException e) {
-            handleError(attribute);
-        } catch (NoSuchMethodException e) {
-            throw new IllegalArgumentException("No getter for " + attribute + " attribute");
-        } catch (InvocationTargetException e) {
-            throw new IllegalArgumentException("Generic type T has no getter for this attribute");
-        }
-        return filtered;
-    }
-
-    private <T> List<T> createFilteredList(Method getter, List<T> list, String value)
-            throws InvocationTargetException, IllegalAccessException {
-        List<T> newList = new ArrayList<>();
-        for (T object : list) {
-            if (getter.invoke(object).equals(value)) {
-                newList.add(object);
-            }
-        }
-        return newList;
-    }
 
     @Override
     public void remove(@NonNull String path, @NonNull String id,
                        @NonNull CompletionListener listener) {
         if (working) {
-            path = path + "/" + id;
             database.remove(path);
-            if(listeners.containsKey(path)) {
-                for(ValueListener pathListener : listeners.get(path)) {
-                    pathListener.onDataChange(null);
-                }
-            }
+            write(path, id, null);
             listener.onComplete(DbError.NONE);
         } else {
             listener.onComplete(DbError.UNKNOWN_ERROR);
         }
     }
 
-    @Override
-    public void readOffers(@NonNull final ValueListener<List<Offer>> listener,
-                           final List<Category> categories) {
-        List<Offer> offers = getList(Database.OFFERS_PATH);
 
-        if (working) {
-            offers = removeOffersWrongCategories(offers, categories);
-            listener.onDataChange(offers);
-        } else {
-            listener.onCancelled(DbError.UNKNOWN_ERROR);
-        }
-    }
-
-    private List<Offer> removeOffersWrongCategories(List<Offer> offers, List<Category> categories) {
-        List<Offer> list = new ArrayList<>();
-        for (Offer o : offers) {
-            if (categories.contains(o.getTag())) {
-                list.add(o);
-            }
-        }
-        return list;
-    }
 
     @Nullable
-    private <T> List<T> getList(@NonNull String path) {
+    protected <T> List<T> getList(@NonNull String path) {
         List<T> list = new ArrayList<>();
         for (Map.Entry<String, Object> entry : database.entrySet()) {
             if (entry.getKey().startsWith(path)) {
@@ -319,6 +176,16 @@ public class FakeDatabase extends Database {
         working = w;
     }
 
+    /**
+     * The working state of the Database.
+     *
+     * @return the working state of the Database, the DataBase will send
+     *         error when working is set at false and will work as
+     *         expected otherwise.
+     */
+    protected boolean isWorking() {
+        return working;
+    }
 
     /**
      * Allow to disable a specific entry for reading. When reading this entry, there will be an
