+++ app/src/main/java/ch/epfl/sweng/swenggolf/offer/ShowOfferActivity.java
@@ -77,6 +77,8 @@ public class ShowOfferActivity extends FragmentConverter {
     private LinearLayout mLayout;
     private View newReaction;
 
+    private ValueListener<Boolean> closingListener;
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
@@ -95,10 +97,27 @@ public class ShowOfferActivity extends FragmentConverter {
             hideReactButton();
             listAnswerAdapter.closeAnswers();
         } else {
+            setClosingListener();
             setAnswerToPost();
         }
         fetchAnswers();
         return inflated;
+    }
+
+    private void setClosingListener() {
+        closingListener = new ValueListener<Boolean>() {
+            @Override
+            public void onDataChange(Boolean value) {
+               if(value != null && value) {
+                   closeOffer();
+               }
+            }
+
+            @Override
+            public void onCancelled(DbError error) { }
+        };
+        Database.getInstance().listen(Database.OFFERS_PATH + "/" + offer.getUuid(),
+                "isClosed", closingListener, Boolean.class);
     }
 
     @Override
@@ -117,7 +136,6 @@ public class ShowOfferActivity extends FragmentConverter {
                 @Override
                 public void onClick(View v) {
                     closeOffer();
-                    closeButton.setVisibility(View.GONE);
                 }
             });
         }
@@ -473,14 +491,16 @@ public class ShowOfferActivity extends FragmentConverter {
      */
     public void closeOffer() {
         offer = new Offer.Builder(offer).setIsClosed(true).build();
+        Database.getInstance().deafen(Database.OFFERS_PATH + "/" + offer.getUuid(),
+                "isClosed", closingListener);
         Database.getInstance().write(Database.OFFERS_PATH, offer.getUuid(), offer);
         hideReactButton();
+        Button closeButton = inflated.findViewById(R.id.close_offer_button);
+        closeButton.setVisibility(View.GONE);
         getActivity().invalidateOptionsMenu();
         listAnswerAdapter.closeAnswers();
         //TODO : add listener
-
     }
-
 
 
 }
