+++ app/src/main/java/ch/epfl/sweng/swenggolf/offer/create/CreatePictureHelper.java
@@ -22,20 +22,22 @@ public final class CreatePictureHelper {
     private static final int STEP_COMPRESS_QUALITY = 5;
     private static final String COMPRESSION_TAG = "compressBitmap";
 
-    private CreatePictureHelper(){}
+    private CreatePictureHelper() {
+    }
 
     /**
      * Creates a compressed version of a file given its path.
-     * @param context the context from which the cache is taken.
+     *
+     * @param context  the context from which the cache is taken.
      * @param filePath the path of the file to be compressed.
      * @param filename the name of the compressed file.
      * @return the Uri where the compressed file is stored.
      */
-     public static Uri compressImageFromPath(Context context, String filePath, String filename) {
+    public static Uri compressImageFromPath(Context context, String filePath, String filename) {
         // First decode with inJustDecodeBounds=true to check dimensions of image
         final BitmapFactory.Options options = new BitmapFactory.Options();
         options.inJustDecodeBounds = true;
-        BitmapFactory.decodeFile(filePath,options);
+        BitmapFactory.decodeFile(filePath, options);
 
         // Calculate inSampleSize(First we are going to resize the image to 800x800 image,
         // in order to not have a big but very low quality image.
@@ -45,9 +47,9 @@ public final class CreatePictureHelper {
 
         // Decode bitmap with inSampleSize set
         options.inJustDecodeBounds = false;
-        options.inPreferredConfig= Bitmap.Config.ARGB_8888;
-        Bitmap bmpPic = BitmapFactory.decodeFile(filePath,options);
-        if(bmpPic == null) {
+        options.inPreferredConfig = Bitmap.Config.ARGB_8888;
+        Bitmap bmpPic = BitmapFactory.decodeFile(filePath, options);
+        if (bmpPic == null) {
             return null;
         }
 
@@ -58,7 +60,7 @@ public final class CreatePictureHelper {
             Toast.makeText(context, "Failed to restore picture original orientation",
                     Toast.LENGTH_SHORT).show();
         }
-        if(tmp != null) {
+        if (tmp != null) {
             bmpPic = tmp;
         }
 
@@ -66,16 +68,15 @@ public final class CreatePictureHelper {
     }
 
 
-
     /**
      * <@see https://stackoverflow.com/questions/39361550/android-resize-image-to-upload-to-server/39363418/>
      */
-    public static Uri compressImageIntoCache(Bitmap bmpPic, Context context, String fileName){
+    public static Uri compressImageIntoCache(Bitmap bmpPic, Context context, String fileName) {
         int targetQuality = computeTargetQuality(bmpPic);
         File temp = compressAndSave(context, fileName, targetQuality, bmpPic);
         //return the path of resized and compressed file
-        if(temp != null) {
-            return FileProvider.getUriForFile(context, "ch.epfl.sweng.swenggolf.fileprovider",temp);
+        if (temp != null) {
+            return FileProvider.getUriForFile(context, "ch.epfl.sweng.swenggolf.fileprovider", temp);
         } else {
             return null;
         }
@@ -84,14 +85,14 @@ public final class CreatePictureHelper {
     private static int computeTargetQuality(Bitmap bmpPic) {
         int compressQuality = MAX_COMPRESS_QUALITY; // quality decreasing by 5 every loop.
         int streamLength;
-        do{
+        do {
             ByteArrayOutputStream bmpStream = new ByteArrayOutputStream();
             Log.d(COMPRESSION_TAG, "Quality: " + compressQuality);
             bmpPic.compress(Bitmap.CompressFormat.JPEG, compressQuality, bmpStream);
             byte[] bmpPicByteArray = bmpStream.toByteArray();
             streamLength = bmpPicByteArray.length;
             compressQuality -= STEP_COMPRESS_QUALITY;
-            Log.d(COMPRESSION_TAG, "Size: " + streamLength/1024+" kb");
+            Log.d(COMPRESSION_TAG, "Size: " + streamLength / 1024 + " kb");
         }
         while (streamLength >= MAX_IMAGE_SIZE);
         return compressQuality;
@@ -102,8 +103,8 @@ public final class CreatePictureHelper {
         File temp = null;
         try {
             //save the resized and compressed file to disk cache
-            Log.d(COMPRESSION_TAG,"cacheDir: "+context.getCacheDir());
-            temp = File.createTempFile(fileName, null,context.getCacheDir());
+            Log.d(COMPRESSION_TAG, "cacheDir: " + context.getCacheDir());
+            temp = File.createTempFile(fileName, null, context.getCacheDir());
             FileOutputStream bmpFile = new FileOutputStream(temp);
             bmpPic.compress(Bitmap.CompressFormat.JPEG, targetQuality, bmpFile);
             bmpFile.flush();
@@ -122,7 +123,7 @@ public final class CreatePictureHelper {
         String debugTag = "MemoryInformation";
         final int height = options.outHeight;
         final int width = options.outWidth;
-        Log.d(debugTag,"image height: "+height+ "---image width: "+ width);
+        Log.d(debugTag, "image height: " + height + "---image width: " + width);
         int inSampleSize = 1;
 
         if (height > reqHeight || width > reqWidth) {
@@ -137,7 +138,7 @@ public final class CreatePictureHelper {
                 inSampleSize *= 2;
             }
         }
-        Log.d(debugTag,"inSampleSize: "+inSampleSize);
+        Log.d(debugTag, "inSampleSize: " + inSampleSize);
         return inSampleSize;
     }
 
@@ -150,7 +151,7 @@ public final class CreatePictureHelper {
         int orientation = ei.getAttributeInt(ExifInterface.TAG_ORIENTATION,
                 ExifInterface.ORIENTATION_UNDEFINED);
         Bitmap rotatedBitmap;
-        switch(orientation) {
+        switch (orientation) {
             case ExifInterface.ORIENTATION_ROTATE_90:
                 rotatedBitmap = rotateImage(photo, 90);
                 break;
