+++ app/src/main/res/drawable/btn_slightly_rounded.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle"
-    android:theme="@style/ToolbarTheme"
-    >
+    android:theme="@style/ToolbarTheme">
 
-    <solid android:color="#f5f5f5"/>
-    <corners android:radius="10dp"/>
+    <solid android:color="#f5f5f5" />
+    <corners android:radius="10dp" />
 </shape>
