+++ app/src/main/res/layout/activity_create_user.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
@@ -24,9 +23,9 @@
         <TextView
             android:id="@+id/presentation"
             android:layout_width="match_parent"
-            android:layout_marginBottom="10dp"
             android:layout_height="wrap_content"
-            android:text="@string/presentation"/>
+            android:layout_marginBottom="10dp"
+            android:text="@string/presentation" />
 
         <TextView
             android:id="@+id/name_field"
@@ -43,17 +42,17 @@
             android:layout_marginLeft="15dp"
             android:layout_marginRight="20dp"
             android:hint="@string/username"
-            android:singleLine="true"/>
+            android:singleLine="true" />
 
         <Button
             android:id="@+id/create_account"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginTop="20dp"
             android:background="@drawable/btn_rounded"
             android:onClick="onClick"
-            android:layout_marginTop="20dp"
             android:text="@string/create_account"
-            android:theme="@style/ButtonTheme"
-            android:layout_gravity="center"/>
+            android:theme="@style/ButtonTheme" />
     </LinearLayout>
 </ScrollView>
