+++ app/src/main/res/layout/activity_leaderboard.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".offer.ListOfferActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/users_recycler_view"
-        android:scrollbars="vertical"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        android:scrollbars="vertical" />
 
     <ImageView
         android:id="@+id/userImage"
@@ -30,8 +30,8 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginLeft="30dp"
         android:layout_marginStart="30dp"
+        android:layout_marginLeft="30dp"
         android:gravity="center"
         app:layout_constraintStart_toStartOf="parent">
