+++ app/src/main/res/layout/activity_notifications.xml
@@ -20,8 +20,8 @@
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/notifications_recycler_view"
-        android:scrollbars="vertical"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:layout_height="match_parent"
+        android:scrollbars="vertical" />
 
 </android.support.constraint.ConstraintLayout>
