+++ app/src/main/res/layout/activity_profile.xml
@@ -19,10 +19,9 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.16">
 
-        <RelativeLayout
-            xmlns:android="http://schemas.android.com/apk/res/android"
+        <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
             android:layout_width="fill_parent"
-            android:layout_height="fill_parent" >
+            android:layout_height="fill_parent">
 
             <ImageView
                 android:id="@+id/ivProfile"
@@ -37,8 +36,8 @@
                 android:id="@+id/ivBadge"
                 android:layout_width="200dp"
                 android:layout_height="220dp"
-                android:src="@drawable/badge00"
-                android:scaleType="fitXY" />
+                android:scaleType="fitXY"
+                android:src="@drawable/badge00" />
         </RelativeLayout>
     </android.support.v7.widget.CardView>
 
@@ -80,8 +79,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
-        android:layout_marginTop="5dp"
         android:layout_marginLeft="16dp"
+        android:layout_marginTop="5dp"
         android:text="@string/user_preferences"
         android:textSize="20sp"
         android:textStyle="bold"
@@ -118,8 +117,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
-        android:layout_marginTop="5dp"
         android:layout_marginLeft="16dp"
+        android:layout_marginTop="5dp"
         android:text="@string/description"
         android:textSize="20sp"
         android:textStyle="bold"
@@ -154,8 +153,8 @@
         android:id="@+id/profileReputation"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="16dp"
         android:layout_marginStart="16dp"
+        android:layout_marginLeft="16dp"
         android:layout_marginTop="20dp"
         android:text="@string/user_reputation"
         android:textSize="20sp"
@@ -167,15 +166,14 @@
         android:id="@+id/points"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="15dp"
         android:layout_marginStart="15dp"
+        android:layout_marginLeft="15dp"
         android:layout_marginTop="25dp"
         android:text="@android:string/unknownName"
         android:textSize="16sp"
         app:layout_constraintLeft_toRightOf="@+id/profileReputation"
         app:layout_constraintStart_toEndOf="@+id/profileReputation"
         app:layout_constraintTop_toBottomOf="@+id/linearLayout2" />
-
 
 
 </android.support.constraint.ConstraintLayout>
