+++ app/src/main/res/layout/activity_show_offer.xml
@@ -20,8 +20,8 @@
                 android:id="@+id/show_offer_author"
                 android:layout_width="wrap_content"
                 android:layout_height="24dp"
-                android:layout_marginLeft="50dp"
                 android:layout_marginStart="50dp"
+                android:layout_marginLeft="50dp"
                 android:layout_marginTop="20dp"
                 android:text="@android:string/untitled"
                 android:textColor="@color/offer_list_author"
@@ -44,11 +44,11 @@
                 android:id="@+id/show_offer_title"
                 android:layout_width="322dp"
                 android:layout_height="wrap_content"
-                android:layout_marginEnd="16dp"
-                android:layout_marginLeft="16dp"
-                android:layout_marginRight="16dp"
                 android:layout_marginStart="16dp"
+                android:layout_marginLeft="16dp"
                 android:layout_marginTop="16dp"
+                android:layout_marginEnd="16dp"
+                android:layout_marginRight="16dp"
                 android:textColor="@color/offer_list_name"
                 android:textSize="24sp"
                 app:layout_constraintEnd_toEndOf="parent"
@@ -59,11 +59,11 @@
                 android:id="@+id/saved_location_offer"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_marginEnd="8dp"
-                android:layout_marginLeft="8dp"
-                android:layout_marginRight="8dp"
                 android:layout_marginStart="8dp"
+                android:layout_marginLeft="8dp"
                 android:layout_marginTop="72dp"
+                android:layout_marginEnd="8dp"
+                android:layout_marginRight="8dp"
                 app:layout_constraintEnd_toStartOf="@+id/show_offer_tag"
                 app:layout_constraintStart_toEndOf="@+id/show_offer_author"
                 app:layout_constraintTop_toTopOf="parent" />
@@ -72,11 +72,11 @@
                 android:id="@+id/show_offer_description"
                 android:layout_width="322dp"
                 android:layout_height="0dp"
-                android:layout_marginEnd="16dp"
-                android:layout_marginLeft="16dp"
-                android:layout_marginRight="16dp"
                 android:layout_marginStart="16dp"
+                android:layout_marginLeft="16dp"
                 android:layout_marginTop="16dp"
+                android:layout_marginEnd="16dp"
+                android:layout_marginRight="16dp"
                 android:textSize="14sp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
@@ -98,9 +98,9 @@
                 android:id="@+id/show_offer_tag"
                 android:layout_width="wrap_content"
                 android:layout_height="24dp"
+                android:layout_marginTop="20dp"
                 android:layout_marginEnd="50dp"
                 android:layout_marginRight="50dp"
-                android:layout_marginTop="20dp"
                 android:text="@android:string/unknownName"
                 android:textColor="@color/offer_list_author"
                 android:textSize="12sp"
@@ -121,10 +121,10 @@
                 android:id="@+id/list_answers"
                 android:layout_width="380dp"
                 android:layout_height="wrap_content"
-                android:layout_marginEnd="8dp"
                 android:layout_marginLeft="16dp"
-                android:layout_marginRight="8dp"
                 android:layout_marginTop="16dp"
+                android:layout_marginEnd="8dp"
+                android:layout_marginRight="8dp"
                 android:orientation="vertical"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
@@ -167,8 +167,8 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_gravity="center_horizontal"
-                        android:layout_marginBottom="10dp"
                         android:layout_marginTop="10dp"
+                        android:layout_marginBottom="10dp"
                         android:background="@drawable/btn_slightly_rounded"
                         android:drawableLeft="@drawable/ic_comment_black_24dp"
                         android:drawablePadding="8dp"
@@ -182,9 +182,9 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_gravity="center_horizontal"
-                        android:layout_marginBottom="10dp"
                         android:layout_marginLeft="4dp"
                         android:layout_marginTop="10dp"
+                        android:layout_marginBottom="10dp"
                         android:layout_toRightOf="@+id/react_button"
                         android:background="@drawable/btn_slightly_rounded"
                         android:drawableLeft="@drawable/ic_close_black_24dp"
