+++ app/src/main/res/layout/leaderboard_list_row.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="130dp"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="?android:attr/selectableItemBackground"
     android:clickable="true"
     android:focusable="true"
@@ -34,15 +34,13 @@
         android:layout_below="@id/user_position" />
 
 
-    <RelativeLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
+    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="100dp"
         android:layout_height="110dp"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
         android:textColor="@color/offer_list_author"
-        app:layout_constraintVertical_bias="0.16"
-        >
+        app:layout_constraintVertical_bias="0.16">
 
         <ImageView
             android:id="@+id/user_image"
@@ -57,8 +55,8 @@
             android:id="@+id/user_badge"
             android:layout_width="100dp"
             android:layout_height="110dp"
-            android:src="@drawable/badge00"
-            android:scaleType="fitXY" />
+            android:scaleType="fitXY"
+            android:src="@drawable/badge00" />
     </RelativeLayout>
