+++ app/src/main/res/layout/notification.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:cardView="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="@dimen/notif_height"
-    xmlns:cardView="http://schemas.android.com/tools"
     android:layout_margin="2dp"
     android:elevation="3dp"
     cardView:cardCornerRadius="2dp">
@@ -16,15 +16,15 @@
             android:layout_width="@dimen/notif_icon_size"
             android:layout_height="@dimen/notif_icon_size"
             android:layout_centerVertical="true"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"/>
+            android:layout_marginLeft="@dimen/activity_horizontal_margin" />
 
         <TextView
             android:id="@+id/notification_text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_centerVertical="true"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
             android:layout_marginStart="@dimen/activity_horizontal_margin"
+            android:layout_marginLeft="@dimen/activity_horizontal_margin"
             android:layout_toEndOf="@id/notification_icon"
             android:layout_toRightOf="@id/notification_icon"
             android:text="@android:string/untitled" />
@@ -33,9 +33,9 @@
             android:id="@+id/clear"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_alignParentTop="true"
             android:layout_alignParentEnd="true"
             android:layout_alignParentRight="true"
-            android:layout_alignParentTop="true"
             android:background="@android:color/transparent"
             android:src="@drawable/ic_clear_black_24dp" />
