+++ app/src/main/res/menu/menu_drawer.xml
@@ -9,8 +9,8 @@
         <item
             android:id="@+id/offers"
             android:icon="@android:drawable/ic_menu_myplaces"
-            android:title="@string/offers"
-            android:onClick="loadShowOffersActivity" />
+            android:onClick="loadShowOffersActivity"
+            android:title="@string/offers" />
         <item
             android:id="@+id/my_offers"
             android:icon="@drawable/baseline_library_books_24"
